refactor: one yt-dlp download implementation in utils; derive CDN extension
Applies the two review findings left open for a maintainer decision, plus the
version bump for release.

1. Adapter coupling (architecture, +3 reviewers). `LoomAdapter` no longer
   instantiates `YtDlpAdapter`. The download body moves to
   `downloadViaYtDlp()` in `src/utils/ytdlp.ts`, which both adapters call.
   That file now owns every yt-dlp concern — findYtDlp, runYtDlp,
   ytdlpCookieArgs, commonArgs, extractYtDlpError, YTDLP_MISSING and the
   download — so the "exactly one download implementation" invariant lives at
   the same seam as runYtDlp instead of inside a sibling adapter.

   Removes all three concrete objections: calling downloadVideo on an adapter
   whose canHandle() returns false for the URL, a second YtDlpAdapter instance
   whose infoInFlight memo didn't span the first, and Loom depending on a
   sibling's whole surface to reuse one method.

   Ordering was as the verifier required: the source guard was widened to scan
   all of src/ (and to fail when it scans nothing) BEFORE this move, so it
   cannot go green-and-empty as a result of it.

2. Loom CDN fallback named its output `.mp4` regardless of what the CDN sent —
   the same "assume the container" assumption that caused #24, one branch over.
   Now derived from Content-Type via extensionFromContentType(), defaulting to
   mp4, with unit tests for the mapping and an end-to-end case asserting a
   video/webm response lands as `<id>.webm`.

3. Version 0.7.0 → 0.7.1 across package.json, src/version.ts and
   .claude-plugin/plugin.json (they must match — CLAUDE.md release checklist).

Docs updated for the new seam: CLAUDE.md (download lives in utils; adapters are
siblings and must not import each other; utils/ytdlp.ts inventory) and
CONTRIBUTING.md (new adapters call downloadViaYtDlp, never another adapter).

Validated after the refactor: 435 unit tests, test:e2e 14 files / 80 tests,
test:smoke, and the real Loom video in both containers — 29 frames / 28 OCR
with and without a system ffmpeg, identical to pre-refactor.
