+++ .claude-plugin/plugin.json
@@ -1,7 +1,7 @@
 {
   "name": "video",
   "description": "Give your agent video input. /video analyzes a URL or local file — transcript, key frames, OCR, timeline — via the mcp-video-analyzer MCP server (auto-registered) or its CLI.",
-  "version": "0.7.0",
+  "version": "0.7.1",
   "author": {
     "name": "Guilherme Matheus"
   },
