+++ CLAUDE.md
@@ -36,9 +36,9 @@ MCP server for video analysis — extracts transcripts, key frames, metadata, OC
 - Use `createTestImage()` from `test/helpers/images.ts` and `FIXTURES_DIR` from `test/helpers/fixtures.ts` — don't redefine in each test file.
 - Use `vitest` with `pool: 'forks'` (required on Windows).
 - Graceful degradation: never throw when partial results are available. Use `warnings[]` array.
-- Three-strategy video download: yt-dlp (primary) → direct HTTP via Loom CDN API (fallback) → headless Chrome screenshots (last resort). `LoomAdapter.downloadVideo` **delegates** its yt-dlp strategy to `YtDlpAdapter.downloadVideo` rather than re-implementing it — there is exactly one yt-dlp download call site in `src/`, and it must stay that way (issue #24 was a second, divergent copy).
+- Three-strategy video download: yt-dlp (primary) → direct HTTP via Loom CDN API (fallback) → headless Chrome screenshots (last resort). **`downloadViaYtDlp()` in `src/utils/ytdlp.ts` is the only yt-dlp download implementation** — both `YtDlpAdapter.downloadVideo` and `LoomAdapter` strategy 1 call it, and it must stay that way (issue #24 was a second, divergent copy that hardcoded `.mp4`). Adapters are siblings behind `IVideoAdapter` and must not depend on each other; shared yt-dlp behaviour belongs in the util.
 - **yt-dlp `-o` templates must never hardcode a container extension.** Always `-o <name>.%(ext)s` plus a `readdir` glob for `<name>.*`; when yt-dlp merges separate DASH video+audio streams it appends the REAL container to whatever template you gave it, so `-o x.mp4` writes `x.mp4.webm` and every `existsSync('x.mp4')` after it silently fails on a download that actually succeeded. Pair it with `--ffmpeg-location ffmpegPath` — without that, yt-dlp can't merge at all in our published image (no system ffmpeg) and leaves the streams separate, so the glob can return the audio-only file. The `%(ext)s` rule is enforced repo-wide by `src/adapters/ytdlp-output-template.test.ts` (which also unit-tests its own detector against the original #24 code); the `--ffmpeg-location` argument is asserted in `src/adapters/ytdlp.adapter.test.ts`.
-- yt-dlp platform URLs (single-video pages only; playlists/channels rejected) route through `YtDlpAdapter`. `findYtDlp()` (positive probe cached per process) + `runYtDlp()` + `ytdlpCookieArgs()` are shared via `src/utils/ytdlp.ts` — always spawn through `runYtDlp` so the bin/prefix pairing can't be forgotten. Missing yt-dlp surfaces as install-hint warnings: adapter `getTranscript`/`getMetadata` throw `YTDLP_MISSING` and every tool handler catches adapter rejections into `warnings[]`; `downloadVideo` must return `null`, never reject (the pipeline calls it without catch) and reports its failure reason via the optional `onWarning` sink. Native captions preferred (uploaded > auto-generated with rolling-window collapse); `[]` from `getTranscript` strictly means "no captions exist" (fetch failures throw) → Whisper fallback.
+- yt-dlp platform URLs (single-video pages only; playlists/channels rejected) route through `YtDlpAdapter`. `src/utils/ytdlp.ts` owns everything yt-dlp: `findYtDlp()` (positive probe cached per process), `runYtDlp()`, `ytdlpCookieArgs()`, `commonArgs()`, `extractYtDlpError()`, `YTDLP_MISSING` and `downloadViaYtDlp()` — always spawn through `runYtDlp` so the bin/prefix pairing can't be forgotten, and add shared flags to `commonArgs()` rather than inlining them at one call site. Missing yt-dlp surfaces as install-hint warnings: adapter `getTranscript`/`getMetadata` throw `YTDLP_MISSING` and every tool handler catches adapter rejections into `warnings[]`; `downloadVideo` must return `null`, never reject (the pipeline calls it without catch) and reports its failure reason via the optional `onWarning` sink. Native captions preferred (uploaded > auto-generated with rolling-window collapse); `[]` from `getTranscript` strictly means "no captions exist" (fetch failures throw) → Whisper fallback.
 - Standard-detail `maxFrames` default is duration-adaptive via `resolveMaxFrames()` in `detail-levels.ts` (~12 for ≤30s up to 60 for >10min). An explicit `maxFrames` always wins and keys the cache separately (`undefined` drops out of the cache key). `get_frames` keeps its fixed default of 20.
 - Silent-audio gate: `transcribeAudio()` probes the track with ffmpeg `volumedetect` (first 2 min) before any Whisper strategy; mean volume ≤ −55dB skips transcription with a warning — an empty transcript on a mute track is content, not a bug.
 - Frame extraction uses bundled `ffmpeg-static` — no system ffmpeg needed.
